Enable correct AnchorText behaviour with spacy>=2.3.0 (#255)
Spacy 2.3.0 moved probability tables outside into a separate package and
the way to access these was slightly changed. This commit enables one to
use either spacy version but crucially relies on installing
spacy_lookups_data even though this is only necessary for spacy>=2.3.0.
Importantly, this data package includes a lot of data files we don't
care about and may grow in the future. We should investigate if there's
a more sensible way of getting the probability scores.