Create `disable_tf2` decorator
This decorator is used for testing functions relying on TF1.x style code
(i.e. CounterFactual, CEM, CFProto). Calling `disable_v2_behavior` at
the top of the test module does not work because at test collection time
it would be executed when we are looking for the mark "not tf1"
therefore ultimately making the tests fail.